fixat try catch

+++ sti-jonathan-filip-service-impl/src/main/java/sti/jonathan/filip/App.java
@@ -7,7 +7,9 @@ import sti.jonathan.filip.domain.Vault;
 import sti.jonathan.filip.service.SchoolService;
 import sti.jonathan.filip.service.impl.SchoolServiceImpl;
 
+import java.util.InputMismatchException;
 import java.util.Scanner;
+import java.util.concurrent.ExecutionException;
 
 public class App {
     private Vault vault;
@@ -44,25 +46,32 @@ public class App {
         boolean loop = true;
         while(loop) {
             System.out.println("Menu \n 1.Add Student \n 2.Register Course \n 3.Remove Course \n 4.Get Student \n 5.Exit");
-            int userChoice = scan.nextInt();
-            //todo lägg till trycatch
-            switch (userChoice) {
-                case 1:
-                    registerStudent();
-                    break;
-                case 2:
-                    registerCourse();
-                    break;
-                case 3:
-                    removeCourse();
-                    break;
-                case 4:
-                    getStudent();
-                    break;
-                case 5:
-                    System.out.println("bye");
-                    loop = false;
+            try {
+                int userChoice = scan.nextInt();
+                //todo lägg till trycatch
+                switch (userChoice) {
+                    case 1:
+                        registerStudent();
+                        break;
+                    case 2:
+                        registerCourse();
+                        break;
+                    case 3:
+                        removeCourse();
+                        break;
+                    case 4:
+                        getStudent();
+                        break;
+                    case 5:
+                        System.out.println("bye");
+                        loop = false;
+                    default:
+                        System.out.println("skriv en siffra mellan 1-5!");
 
+                }
+            }catch (InputMismatchException e){
+                scan.next();
+                System.out.println("Skriv en siffra tack!");
             }
         }
     }
@@ -84,18 +93,38 @@ public class App {
         System.out.println("Vilken elev vill du ta bort en kurs ifran? Skriv personnummer");
         String person = scan.nextLine();
         Student student = schoolService.getStudent(person);
-        System.out.println("vilken kurs vill du ta bort? \n 1.java \n 2.html \n 3.css");
-        int course = scan.nextInt();
-        switch(course){
-            case 1:
-                schoolService.removeCourse(student, schoolService.getCourse("java"));
-                break;
-            case 2:
-                schoolService.removeCourse(student, schoolService.getCourse("html"));
-                break;
-            case 3:
-                schoolService.removeCourse(student, schoolService.getCourse("css"));
-                break;
+        if (student == null){
+            System.out.println("Finns ingen person med detta personnummer");
+        }
+        else {
+
+
+            boolean catcher = true;
+            System.out.println("vilken kurs vill du ta bort? \n 1.java \n 2.html \n 3.css");
+            while (catcher) {
+                try {
+                    int course = scan.nextInt();
+                    switch (course) {
+                        case 1:
+                            schoolService.removeCourse(student, schoolService.getCourse("java"));
+                            catcher = false;
+                            break;
+                        case 2:
+                            schoolService.removeCourse(student, schoolService.getCourse("html"));
+                            catcher = false;
+                            break;
+                        case 3:
+                            schoolService.removeCourse(student, schoolService.getCourse("css"));
+                            catcher = false;
+                            break;
+                        default:
+                            System.out.println("ska vara en siffra mellan 1-3!");
+                    }
+                } catch (InputMismatchException e) {
+                    scan.next();
+                    System.out.println("skriv en siffra mellan 1-3!");
+                }
+            }
         }
     }
 
@@ -103,23 +132,43 @@ public class App {
         Scanner scan = new Scanner(System.in);
         System.out.println("Vilken student ska du registrera en kurs till?\n skriv personnummer");
         String person =scan.nextLine();
-        // todo kolla så att den inte blir null innan vi försöker lägga till kurser
         Student student = schoolService.getStudent(person);
-        System.out.println("vilken kurs vill du registrera?\n 1.html \n 2.css \n 3.java");
-        int userChoice = scan.nextInt();
-        //todo add try catch ifall man skriver in bokstäver
-        switch (userChoice){
-            case 1:
-                schoolService.registerCourse(student, schoolService.getCourse("html"));
-                break;
-            case 2:
-                schoolService.registerCourse(student, schoolService.getCourse("css"));
-                break;
-            case 3:
-                schoolService.registerCourse(student, schoolService.getCourse("java"));
+        if (student == null){
+            System.out.println("Finns ingen person med detta personnummer");
         }
-    }
+        else {
 
+            System.out.println("vilken kurs vill du registrera?\n 1.html \n 2.css \n 3.java");
+
+            boolean catcher = true;
+            while (catcher) {
+                try {
+                    int userChoice = scan.nextInt();
+                    switch (userChoice) {
+                        case 1:
+                            schoolService.registerCourse(student, schoolService.getCourse("html"));
+                            catcher = false;
+                            break;
+                        case 2:
+                            schoolService.registerCourse(student, schoolService.getCourse("css"));
+                            catcher = false;
+                            break;
+                        case 3:
+                            schoolService.registerCourse(student, schoolService.getCourse("java"));
+                            catcher = false;
+                        default:
+                            System.out.println("ska vara en siffra mellan 1-3!");
+                    }
+                } catch (InputMismatchException e) {
+                    scan.next();
+                    System.out.println("skriv en siffra mellan 1-3!");
+                }
+            }
+        }
+
+
+    }
+//todo se till satt personnummer bara är siffror och är 10 char long
     private void registerStudent() {
         Scanner scan = new Scanner(System.in);
         System.out.println("First name: ");
